Move test_file_writer to src/common
Modify noh with viz
