+++ src/noh/noh.cpp
@@ -11,8 +11,9 @@
 #include "cstone/domain/domain.hpp"
 #include "sphexa.hpp"
 #include "sph/findNeighborsSfc.hpp"
-#include "test_file_writer.hpp"
+#include "common/test_file_writer.hpp"
 #include "propagator.hpp"
+#include "insitu_viz.h"
 
 #include "NohDataGenerator.hpp"
 
@@ -20,25 +21,11 @@ using namespace cstone;
 using namespace sphexa;
 using namespace sphexa::sph;
 
-#ifdef SPH_EXA_USE_CATALYST2
-#include "CatalystAdaptor.h"
-#endif
-
-#ifdef SPH_EXA_USE_ASCENT
-#include "AscentAdaptor.h"
-#endif
-
 void printHelp(char* binName, int rank);
 
 int main(int argc, char** argv)
 {
-    const int rank = initAndGetRankId();
-
-#ifdef SPH_EXA_USE_CATALYST2
-    CatalystAdaptor::Initialize(argc, argv);
-    std::cout << "CatalystInitialize\n";
-#endif
-
+    const int       rank = initAndGetRankId();
     const ArgParser parser(argc, argv);
 
     if (parser.exists("-h") || parser.exists("--h") || parser.exists("-help") || parser.exists("--help"))
@@ -92,10 +79,8 @@ int main(int argc, char** argv)
 
     if (d.rank == 0) std::cout << "Domain synchronized, nLocalParticles " << d.x.size() << std::endl;
 
-#ifdef SPH_EXA_USE_ASCENT
-    AscentAdaptor::Initialize(d, domain.startIndex());
-    std::cout << "AscentInitialize\n";
-#endif
+    viz::init_catalyst(argc, argv);
+    viz::init_ascent(d, domain.startIndex());
 
     const size_t nTasks = 64;
     const size_t ngmax  = 150;
@@ -125,27 +110,13 @@ int main(int argc, char** argv)
 #endif
         }
 
-#ifdef SPH_EXA_USE_CATALYST2
-        CatalystAdaptor::Execute(d, domain.startIndex(), domain.endIndex());
-#endif
-
-#ifdef SPH_EXA_USE_ASCENT
-        if (d.iteration % 5 == 0) { AscentAdaptor::Execute(d, domain.startIndex(), domain.endIndex()); }
-#endif
+        if (d.iteration % 5 == 0) { viz::execute(d, domain.startIndex(), domain.endIndex()); }
     }
-
+viz
     totalTimer.step("Total execution time of " + std::to_string(maxStep) + " iterations of Noh");
 
     constantsFile.close();
-
-#ifdef SPH_EXA_USE_CATALYST2
-    CatalystAdaptor::Finalize();
-#endif
-
-#ifdef SPH_EXA_USE_ASCENT
-    AscentAdaptor::Finalize();
-#endif
-
+    viz::finalize();
     return exitSuccess();
 }
